requirements.txt
django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.5
psycopg2-binary==2.9.9
celery==5.3.1
redis==4.6.0
django-redis==5.4.0
drf-yasg==1.21.7
django-simple-history==3.4.0
python-dotenv==1.0.0
psutil==5.9.8
coverage==7.3.4
factory-boy==3.3.0
gunicorn==21.2.0
whitenoise==6.6.0
django-extensions==3.2.3